repo_bash_autocompletion: exclude branch if it includes 'master'
This is the situation we're trying to prevent:

(cr) (master stable/(8449a33...)) me@pc0 ~/trunk/src/scripts $

We instead want the old behavior, which excludes the branch name(s) if we're just
in the master branch.

Checking if the branch remotes contains master seems to maintain present
behavior reasonably closely, now that people will be using the "stable"
remote ref as well, with the new sync-to-green feature.
https://groups.google.com/a/google.com/forum/#!topic/chromeos-eng/4t9eczZQDfo

BUG=chromium:985030
TEST=local run

Exempt-From-Owner-Approval: no one has touched this file in years
Change-Id: I2ad7a286fdbd8ab368f033e2d9f1f55f297f83dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1776681
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>