deploy_termina: switch to termina_build_image.py
Change the deploy_termina script to use the Python implementation of
termina_build_image.py as part of deprecating the shell version.

The new script needs to run as root, so a few `sudo` calls are added to
make this seamless when run from within cros_sdk.

termina_build_image.py doesn't need the --arch option, so we can delete
the $arch variable entirely.

BUG=chromium:999698
TEST=deploy_termina -d $DUT

Change-Id: If554fa45cc708141be7a994a6f82b40c5d1c2f12
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1779140
Reviewed-by: Stephen Barber <smbarber@chromium.org>