nebraska: Allow passing no_update=true in the query string
Sometimes we need to pass noupdate=true to the client even if we have an
update payload ready. We already could do this in nebraska for getting
responses through python API calls, but there is another use case of it
in the autoupdate_EndToEndTest autotest (with nebraska running on the
DUT) where we do an update check right after an update+reboot. This
update check happens on the same update metadata directory (which
includes the payload itself) as the previous update, so nebraska
responds with an actual payload address. We need a way to send
noupdate=true to the client on fdemand. There are other ways to do this,
like passing an argument when starting the nebraska or having separate
directories for payload properties file and payload itself but, the
approach in the CL is the most scalable one.

BUG=chromium:920404
TEST=./nebraska_unittest
TEST=./nebraska.py and send a request with no_update=true in the URL.

Change-Id: Id82fcde4c1e303cafd262e7c77eafd355036c9fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1785138
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>