nebraska: Use update and requestlog path in URL
Currently we can call nebraska URL with any random path like:
- http://localhost/some/random/path/&query=true

Not that it is causing any problem, but there is a use case for
supporting multiple pre-defined paths (same as what devserver.py does
using cherrypy). This use case involves requesting the nebraska to
return the list of requests (came from update_engine) in a json
format. This list is used in the autotests to test the flow of the
update.

As of this CL, a POST request for update/install can only go through the
"update" path such as:
- http://localhost:ip/update/&query=true

A GET request for logs can go through the "requestlog" path such as:
- http://localhost:ip/requestlog

Also the auto tests require a timestamp value (when the request was
received) that is added to the requestlog.

BUG=chromium:920404
TEST=./nebraska_unittest
TEST=sudo ./nebraska.py --update-metadata ~/trunk/images/test-payload/ \
	  --port 10000
and then send this requests to it to get empty log:
$ curl -X GET http://127.0.0.1:10000/requestlog
[]

Then send a post request to get actual response:
$ curl -X POST -d @request.xml http://127.0.0.1:10000/update

Then the request for log:
$ curl -X GET http://127.0.0.1:10000/requestlog
[{"event_result": "1", "event_type": "54", "previous_version": "12463.0.2019_09_04_0948", "track": "stable-channel", "timestamp": "2019-09-11 11:54:22", "ping": false, "request_type": 2, "version": "12365.0.0", "board": "reef", "appid": "{8396029B-FCEF-9EEC-C684-3BCB8E3E9429}", "hardware_class": "REEF F2B-B4B-22K-N3S-K8L-A2V DEV", "delta_okay": true}]

Also for failed cases I sent requests like:
$ curl -X GET http://127.0.0.1:10000/invalid-path
<head>
<title>Error response</title>
</head>
<body>
<h1>Error response</h1>
<p>Error code 400.
<p>Message: The requested path "invalid-path" was not found!.
<p>Error code explanation: 400 = Bad request syntax or unsupported method.
</body>

$ curl -X POST http://127.0.0.1:10000/invalid-path
<head>
<title>Error response</title>
</head>
<body>
<h1>Error response</h1>
<p>Error code 400.
<p>Message: Invalid request (header)..
<p>Error code explanation: 400 = Bad request syntax or unsupported method.
</body>

Change-Id: I327a4e51fc5391d882ab22953c8e4c6c4c3d092a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1787982
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>