nebraska: Add default log file
We almost always need to dump the nebraska log somewhere but currently
the default behavior is logging to stdout. This CL changes this so by
default we log in /tmp/nebraska.log and if anyone wanted to see the log
in stdout, then can pass '--log-file stdout' to nebraska.py

Also disable logging in unit test so it doesn't pollute the unit test
output.

BUG=chromium:920404
TEST=./nebraska.py # logs into /tmp/nebraska
TEST=./nebraska.py --log-file stdout # logs to stdout

Change-Id: I70603f23cbcd3afabe4af59e4b78d28b47be61be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1797193
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>