nebraska: Properly handle non-sudo runs
Currently if the nebraska is run on a host machine as non-sudo it
fails to write its pid/port files into /run/nebraska. And then it exists
the main thread on exception, but since we have already started the
server thread, the process gets stucked and we can't really exist it by
other means than pressing ctrl + z to minimize the zombie process, find
its pid and kill -9 it. We should handle these exceptions before we
start the server.

This CL moves the logic of these two files into the NebraskaServer
itself so we can catch (and exit) the exceptions before starting the
server thread. In addtion this changes the default value for the runtime
root to None when the NebraskaServer is used as a library. The reason is
whoever stars the NebraskaServer as a library does not really need to
know the PID and it can get the port number using GetPort() funtion of
NebraskaServer. This means we won't write into port and pid files
either. This is the use case which will replace nano_omaha_devserver.py.

This chnage also makes the main functions logic more compact and
concentrated.

BUG=chromium:920404
TEST=./nebraska_unittest
TEST=./nebraska.py #non-sudo. fails but exists gracefully.
Logging to /tmp/nebraska.log
Permission error: You need to run the script as root/sudo or change the --runtime-root to point to a non-root accessible location.
Traceback (most recent call last):
  File "./nebraska.py", line 993, in <module>
    sys.exit(main(sys.argv))
  File "./nebraska.py", line 985, in main
    nebraska_server.Start()
  File "./nebraska.py", line 874, in Start
    with open(self._port_file, 'w') as port_file:
IOError: [Errno 13] Permission denied: '/run/nebraska/port'

TEST=sudo ./nebraska.py

Change-Id: I8ba6f39e07e779d7874ba6c8406a17f33af98846
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1797194
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>