fromupstream: Support picking patches with message-id
We're moving towards using message-ids for tracking patches to the linux
kernel. Instead of requiring users to find the patchwork ID from the
website/UI, support deriving that ID based on the message-id passed in
as an argument. The same semantics as patchwork picks apply here, so
reuse a bunch of that code and hardcode a few different locations to
look for patchwork instances that can supply the patch in mbox format.

BUG=chromium:1008546
TEST=run with valid message id and invalid message id for a patch on
lkml

Change-Id: Ie349e890221b1a421938faf837d4b0b9e6e39937
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1828169
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>