start_devserver: drop common.sh usage
We were using common.sh for restart_in_chroot_if_needed, except that
we already tried to load a chroot-specific script before the call,
and then exited if we couldn't.  No one seems to care, so lets just
drop the logic entirely and run the devserver directly.  If we aren't
in the chroot, then the end result is still going to be the same.

BUG=None
TEST=CQ passes

Change-Id: I257340b9aea3f6f13ad8ef4f43208cf3d36ac713
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1843273
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>