cros_update: Fix starting cros_update in devserver
CL:1865147 and CL:1871765 did not properly test the use case of
cros_update through autoupdate_Endtoendtest and it is now failing those
tests (we haven't had a devserver push, but if we had, those tests would
be failing now.). The reason was that I moved the cros_update to
chromite and updated devserver to use the new code, but when I got
feedback that we need to fix the cros_update in its original repo first,
I forgot to fix the caller of cros_update and test it with endtoend
tests. So this happened.

This CL moves back calling cros_update.py to what it was before except
that we don't pass shell=True (default is False) anymore since we're
passing the arguments as a list I guess (it doesn't work with
shell=True).

BUG=chromium:1018237, chromium:1003986
TEST=autoupdate_EndToEndTest (look in the control file):
test_that 100.90.30.183 autoupdate_EndToEndTest --args="target_release=12630.0.0 target_payload_uri='gs://chromeos-releases/canary-channel/reef/12630.0.0/payloads/chromeos_12630.0.0_reef_canary-channel_full_test.bin-gvsgentbhe3weyrvxdf7m2jmagoju4pw'"

Change-Id: I1223a5cbc0d22078beaa31aad3486501152efe56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1888436
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>