Revert "devserver: Stop patching devserver internals for port=0"
This reverts commit 0ee73f408b6fa66b1f7defadb1073ff051567010.

In addition, this makes hacking --port=0 on cherrypy conditional on the
version of the cherrypy. The reason we need this is that our devservers
run on older linux and their cherrypy can't be upreved easily. But for
Chrome OS we have already upreved cherrypy and we don't need this
hacking.

There are also other pylint errors fixed.

BUG=chromium:1018237
TEST=devserver_integration_test

Change-Id: If65b75deb565ca78353803c8eab3f7cfc3851e5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1914538
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>