variant: use a list of stages in the yaml to control flow
Rather than have a single fixed map of current stage to next stage,
assign a list of stages based on the baseboard name, and use that
list to control the flow of the program.

BUG=b:146646594
TEST=`./new_variant.py --board=hatch --variant=sushi`
When the script stops, examine ~/.new_variant.yaml
Look for a stage_list entry with several elements in a list.

Clean up by using repo abandon for the branches that new_variant
created, and rm ~/.new_variant.yaml

Change-Id: I7adc2823079d9fb9c26edfa5b864de4d7bb015b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1990410
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>