variant: use baseboard name for all packages
There were several package names that still included "hatch" as
a fixed string; replaced these with the baseboard name. Also
update the package name for the FSP to use the appropriate
chipset.

BUG=b:146646594
TEST=Run the script to create variants of hatch and volteer
baseboards. Ensure that it uses the correct directories for each
of them.

Change-Id: I9551789cece6b354cdb4378d342214abb577266c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1991541
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>