devserver.py: Handle query strings for critical updates
Currently, the only way for devserver to return a response indicating a
critical update is to spawn it with '--critical_update' flag. This is
not an ideal implementation as the autotests that require a critical
update response will need to spawn a new devserver (probably along with
the lab devservers) to achieve this. A much easier approach is to pass
"critical_update=True" query string to the update URL passed to the
update_engine. This way the devserver/nebraska would know that response
to this specific request should be critical.

BUG=chromium:1004487
TEST=devserver_integration_test.py
TEST=manually started devserver.py and send a request to this URL:
http://127.0.0.1:8080/update/?critical_update=True

Change-Id: Ie9908ce7ae83d5ec29534b1e2cba6e60bf242dbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1996147
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>