nebraska: Add search query support for remaining of ResponseProperties
Other members of ResponseProperties need to be accepted in the query
strings so we can use them in autotests.

Also added a path for health_check as we do it already in cros flash and
we will need it in autotets.

Added respective unittests too.

BUG=chromium:999047
TEST=nebraska_unittest.py
TEST=test_that 100.90.30.183 autoupdate_EOL

Change-Id: I0413ae3be359acb42cd06d20dc0658c6ebe31222
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2004054
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>