add_variant_to_yaml: bump version number
Bump the version number to 1.1.0 (minor version adds functionality
in a backward-compatible way). This should have been done when
support for Volteer was added.

BUG=None
TEST=run add_variant_to_yaml.sh and verify that it uses the new
version number in the commit message it creates.

Exempt-From-Owners-Approval: This is a simple change in a text
string that is only used in a commit message.

Change-Id: I08ff633eb4d53d7a63834c6ac947fa08fa9c0108
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2009026
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>