new_variant: support private repos
Some reference designs do not yet have all of their code in public
repos, but we still need to be able to create new variant boards.
Add variables to reference board data for where coreboot source and
coreboot config are located. Use these variables to launch scripts
in the correct directory, or to point to those directories with
an environment variables that the launched script will use.

BUG=b:147483696
TEST=`./new_variant.py --board=hatch --variant=sushi`
Verify that the commit for config.sushi is in
third_party/chromiumos-overlay/sys-boot/coreboot/files/configs

`./new_variant.py --board=volteer --variant=ripto`
Verify that the commit for config.ripto is in
private-overlays/chipset-tgl-private/sys-boot/coreboot/files/configs

Change-Id: Ide00562ca7308f29c1732afbd234fc504b933ecd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2025291
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jacob Rasmussen <jacobraz@google.com>