new_variant: use coreboot-zork directory
The zork project requires that the coreboot files are in coreboot-zork.
Making a symbolic link from third_party/coreboot to
third_party/coreboot-zork does not suffice, so adjust the path in
the trembyle configuration data.

BUG=b:148161697
TEST=`./new_variant.py --board=trembyle --variant=dalboz`
Verify that new files are created in third_party/coreboot-zork and
that the emerge command for coreboot succeeds.

Change-Id: I13b90ea69a5b561929a8193ba5f53a12b9076bfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2026260
Reviewed-by: Eric Peers <epeers@google.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>