create_initial_ec_image: update copyright year in copied files
When create_initial_ec_image.sh copies the reference board files
to start an EC build for a new variant, it should update the
copyright year in each file so that we don't get a warning from
the presubmit hook that the copyright year is out of date.

BUG=b:148630125
TEST=`./create_initial_ec_image.sh hatch sushi`
Examine the commit and observe that the copyright years are the
current year, not the old one in the hatch/* files.

Change-Id: Ie84a9294d0ed6e8df71efc6fe1a14db1d288aeed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2033496
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>