new_variant: do not check for dev.bin and net.bin
If a board does not build depthcharge yet, we won't get an
image-${VARIANT}.net.bin out of the build, even though the build
was successful. Stop checking for this file to determine the success
of the build. The .dev.bin is rarely used, so remove that, too.

BUG=b:149040756
TEST=`./new_variant.py --board=trembyle --variant=grue`
Verify that the emerge command succeeds, even though image-grue.net.bin
doesn't exist when the emerge is done.

Change-Id: I46a58cff63a1bf89992d90a90b28df01b73cd14c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2044611
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Auto-Submit: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Martin Roth <martinroth@google.com>