new_variant: copy cras config for new variant
The Zork boards need to have their own cras config as part of the
initial build, and these should be copied from the reference
board's cras config. Add a new script to copy the subdir and make
a commit out of it, then call that from new_variant.py.

BUG=b:149048513
TEST=`./copy_cras_config.sh zork trembyle grue`
Look in overlays/overlay-zork/chromeos-base/chromeos-bsp-zork/files/cras-config
for a commit that adds files in Grue and that Grue/* is a copy of Trembyle/*

Then test integration with new_variant:
`./new_variant.py --board=trembyle --variant=grue`
Check for the commit in the cras-config folder (listed above) and see
that the build is complete.

Change-Id: Ifc117f9a4cd74b5bfde38d27d54b22215ee9f0ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2044883
Auto-Submit: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Martin Roth <martinroth@google.com>
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>