trembyle: remove unnecessary packages from build
We don't need to workon or emerge libpayload or depthchange, and
we don't need to workon vboot_reference to get the chromeos-bootimage
for a Zork board, so remove them from the list of packages.

BUG=b:148161697
TEST=`./new_variant.py --board=trembyle --variant=grue`
Verify that the program - including the emerge at the end - completes
successfully.

Change-Id: I0e3af57443dbff838ea53c829cae6115f698ea1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2044974
Auto-Submit: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Martin Roth <martinroth@google.com>
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>