new_variant: upload commits to gerrit
Change the keys that are used to store commits in the yaml to use
the strings from step_names, for consistency. Use those names to
have a list of commits that will be uploaded to gerrit.

When uploading commits, search both internal and external gerrit
for the change-id, and if found, record the gerrit instance and the
CL number in the yaml file. If the change-id is not found, launch
repo upload to upload the CL and then after it's uploaded, go back
and search for it again by change-id and save the instance and CL
in the yaml.

BUG=b:149508647
TEST=`./new_variant.py --board=trembyle --variant=grue`
When the program finishes, you will have one CL in internal gerrit
for the changes to model.yaml, and four CLs in external gerrit for
coreboot-zork, the coreboot config, the cras config, and the EC
image. All of these will be marked WIP.

Change-Id: I198b48ab07e6d36ca3eb6b568f51e36cb19b992b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2067068
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jacob Rasmussen <jacobraz@google.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>