new_variant: ask user to push to coreboot
For any commit that needs to be pushed to coreboot, print a message
asking the user to push it to review.coreboot.org. When the program
resumes, check if the changes have been pushed, and if not, ask
again.

See the bug for information about the issues that prevent this from
being automated right now.

BUG=b:150130960
TEST=`./new_variant.py --board=hatch --variant=sushi`
When the program stops and asks you to generate the fitimage, do
that and then `./new_variant.py --continue`
The program stops again to ask you to pushi a CL to coreboot. Do
that, and then `./new_variant.py --continue --verbose`. Look at the
messages and see that it locates the CL on review.coreboot.org,
and then continues to upload the other CLs to internal and external
gerrit.

Change-Id: I046d1158b784cb0e584dd800ca51d7303df3c4ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2072876
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Jacob Rasmussen <jacobraz@google.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>