new_variant: find the upstreamed coreboot CL
After a coreboot patch is merged, it will get upstreamed to the
chromiumos tree as part of a batch. The change-id from the coreboot
gerrit instance will be preserved as original-change-id, so search
for that to find the upstreamed CL in chromiumos gerrit.

Change the messages about doing things outside of new_variant.py
(like generating the fitimage, pushing to coreboot, or waiting for
upstream) from info to error, because these are technically error
conditions that prevent the program from continuing.

BUG=b:150403065
TEST=Check operation when the coreboot CL hasn't been pushed yet,
when it hasn't been upstreamed yet, and when it has been upstreamed.
See testdata/README.md for detailed instructions.

Change-Id: I9f701664d849197cf183fc1fb46f7523095c359c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2079691
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: Martin Roth <martinroth@google.com>
Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>