nebraska: Move query parsing logic into a single function
Currently, the use case in autoupdate.py for kwargs that it gets is not
quite correct as the dictionary passed to nebraska.py is not correctly
formatted. This CL fixes the issue by converting the given kwargs to a
proper dictionary.

BUG=None
TEST=nebraska_unittest, autoupdate_unittest

Change-Id: Ia8afb391511380df41bf09bcb5811917c8b1eef2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2080619
Tested-by: Amin Hassani <ahassani@chromium.org>
Auto-Submit: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>