nebraska: Re-write unittests
The current unittests are not ideal. They do tests implementation rather
than behavior and they have become very hard to modify and
maintain. This CL mainly re-writes the unittests so someone can easily
generate an XML request, get an XML response and look into the response
for properties they are looking for. This case we don't really care the
implementation details of nebraska as long as the behavior is
sufficient. This also removed a ton of duplicate unittest and ones that
were not really testing anything useful.

In addition I figured out these problems in the nebraska code while
doing this:
- is_platform is not used anywhere. removed it + unittests
- introduced Request.has_update_check. This can define whether an App
  request is for platform or not in regard to installs.
- For installs, if there were a payload for the platform app, we would
  return it, but we should not do that.
- On installs, the platform app response should not contain update
  check.
- do utf-8 decoding of sha256_hex a bit sooner.
- Do not return unknowapplication error anymore. We don't need this.

BUG=chromium:914936
TEST=nebraska_unittest
TEST=cros flash
TEST=passes tast-dlc in CQ

Change-Id: Ic1e95c2330ecc2b66636e05fa0d222b60dcdc567
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2088761
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>