nebraska: Allow full_payload=True
Currently, nebraska chooses a delta payload over a full payload if
properties file for both exists in the same directory. However, some
tests require a full payload, but sometimes the devserver cache
directory contains for the full and delta payload. So we need to
specifically tell the nebraska to choose the full payload.

BUG=chromium:1061507
TEST=./nebraska_unittest.py

Change-Id: I3483b90da5ccd8d0dc437810757661b9df6e8f6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2105930
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>