volteer: use default directory for coreboot config
new_variant.py and create_coreboot_config.sh allow specifying a
directory for the coreboot config file, or accepting the default.
With volteer moving to the public coreboot repos, update the
directory for coreboot config to use the default (which is the
public repo).

BUG=b:147483696
TEST=create a volteer variant and see that the updated config
file is in the public repo:
third_party/chromiumos-overlay/sys-boot/coreboot/files/configs

Change-Id: I7e57d0c3fe325b9beb79c7585d4d62ba8d895800
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2107658
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>