fromupstream: add support for non kernel.org archives
The 'also from' messages are not correct in cases where lore.kernel.org
doesn't have an appropriate archive. This is the case for hostap
project:

http://lists.infradead.org/pipermail/hostap/
https://patchwork.ozlabs.org/project/hostap/list/

It's not strictly a kernel project, so it probably doesn't meet the
kernel.org requirements:

https://korg.wiki.kernel.org/userdoc/lore

It is, however, available via the https://marc.info/?i=${MESSAGE_ID}
redirector.

Let's probe for non-404 responses, and iterate through a few different
archive sources. Note that public-inbox.org currently gives a nice list
of possible other redirects, so even if nothing is directly found, let's
still encode the MessageId in a public-inbox URL, with the hope that
maybe it can still be useful in finding a good archive.

Note that py3 upgrades (https://crrev.com/c/1963397) overlooked the fact
that the new urllib raises HTTPError for non-200 erro codes, so the
existing 'opener.get_code() != 200' check is obviated. Factor out a
_try_urlopen() to make this a little nicer to use.

BUG=none
TEST=`fromupstream.py -b= -t= pw://hostap/1255841`, where hostap points
     at ozlabs; URL output:
       (am from https://patchwork.ozlabs.org/patch/1255841/)
       (also found at https://marc.info/?i=20200316211106.131858-1-matthewmwang@chromium.org)

Change-Id: I00b0ed3c52580cfc038b1ba3e9ee3d2cb62d16a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/2108789
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>